# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:8080/api
    variables: {}
paths:
...
  /auth/profile/{id}:
    get:
      tags:
        - Auth
      summary: Get User Profile by ID
      operationId: GetUserProfile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
...
  /auth/delete/{email}:
    delete:
      tags:
        - Auth
      summary: Delete User by Email
      operationId: DeleteUserByEmail
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
...
components:
  schemas:
...
    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...